refactor(desktop): split workspacessidebar agents tree into parts
