+++ apps/desktop/src/features/workspace/components/WorkspacesSidebar/index.tsx
@@ -1,107 +1,30 @@
-import { Fragment, useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { useShallow } from 'zustand/react/shallow';
-import { createPortal } from 'react-dom';
-import { Button, Dialog, Divider, Popover, ScrollArea, SectionHeader, cn } from '@goodboy/ui';
-import {
-  AlertTriangle,
-  ArrowRight,
-  Ban,
-  Check,
-  CheckCheck,
-  ChevronDown,
-  ChevronRight,
-  ChevronUp,
-  Clock,
-  DollarSign,
-  Gauge,
-  GitCommit,
-  HelpCircle,
-  Layers,
-  Link2,
-  Loader2,
-  MessageSquareReply,
-  Moon,
-  PanelLeftClose,
-  Pause,
-  Play,
-  Plus,
-  Settings,
-  Sun,
-  Trash2,
-  X,
-  Zap,
-  ZapOff,
-} from 'lucide-react';
+import { useCallback, useState } from 'react';
+import { Divider, ScrollArea } from '@goodboy/ui';
+import { DollarSign, HelpCircle, Moon, PanelLeftClose, Settings, Sun } from 'lucide-react';
 import { GuideDialog } from '../../../settings/components/GuideDialog';
 import { NotificationCenter } from '../../../../features/notifications/components/NotificationCenter';
-import { ScriptsSection } from '../../../scripts/components/ScriptsSection';
 import { WORKSPACE_FEATURES } from '../../../../shared/lib/features';
-import { DogMascot } from '../../../../shared/components/DogMascot';
 import { OnboardingChip } from '../../../onboarding/OnboardingCard';
-import type {
-  Agent,
-  AgentId,
-  ProviderRunId,
-  Session,
-  SessionId,
-  TelemetryRecord,
-  TurnState,
-  Workflow,
-  WorkflowRun,
-  WorkflowRunId,
-  WorkspaceId,
-} from '@goodboy/types';
+import type { Session, SessionId, WorkspaceId } from '@goodboy/types';
 import {
   EMPTY_ARRAY,
-  agentHasUnread,
   useAppStore,
   useCurrentSession,
   useCurrentWorkspace,
-  useSessionLoading,
-  useSessionOpenQuestions,
   useSessions,
-  useWorkspaces,
 } from '../../../../store';
-import { pickNextWorkflowStep } from '../../../../features/workflows/components/WorkflowNextStepCta';
-import { workflowRunHasOpenQuestions } from '../../../../features/context/openQuestionsGate';
-import {
-  computeLatestTelemetryByAgentId,
-  formatCost,
-  formatTokens,
-} from '../../../../features/session/agent-row-format';
-import { PROVIDER_CAPABILITIES, WORKFLOW_LIBRARY, getModelProvider } from '@goodboy/core';
-import {
-  AGENT_KIND_DEFAULTS,
-  AGENT_KIND_META,
-  AGENT_KIND_ORDER,
-  AGENT_KIND_PALETTE,
-  type AgentKind,
-  inferAgentKindFromName,
-  resolveAgentKind,
-} from '../../../../features/session/agent-kind';
-import { AgentKindChip } from '../../../../features/session/components/AgentKindChip';
-import {
-  AgentMetricsBlock,
-  type AgentAggregate,
-} from '../../../../features/session/components/AgentMetricsBlock';
-import { formatError } from '../../../../shared/lib/errors';
 import { useThemeStore } from '../../../../shared/lib/theme';
 import { WorkspaceHeader } from '../WorkspaceHeader';
 import { WorkspaceLinkDialog } from '../WorkspaceLinkDialog';
 import { SessionActivityBar } from '../SessionActivityBar';
 import { SessionDetailPanel, SessionMetaFooter } from '../SessionDetailPanel';
-import { formatRelativeDuration } from '../../../../shared/utils/relativeDate';
-import { useNow } from '../../../../shared/hooks/useNow';
-import { openUrl } from '../../../../shared/lib/editor';
 import { FOOTER_ICON_BTN } from './lib';
 import { CollapsedSidebarRail } from './parts/CollapsedSidebarRail';
 import { QuickActionsRow } from './parts/QuickActionsRow';
-import { SectionToggle } from './parts/SectionToggle';
 import { SidebarLogo } from './parts/SidebarLogo';
 import { SidebarDetailHint } from './parts/SidebarDetailHint';
 import { NoWorkspaceEmpty } from './parts/NoWorkspaceEmpty';
-import { PlanReadySuggestion } from './parts/PlanReadySuggestion';
-import { ProviderGlyph } from './parts/ProviderGlyph';
+import { AgentsSection } from './parts/AgentsSection';
 
 type WorkspacesSidebarProps = {
   onOpenSettings: () => void;
@@ -316,1976 +239,3 @@ export const WorkspacesSidebar = ({
     </div>
   );
 };
-
-function workflowKindName(workflow: Workflow): string {
-  const raw = workflow.name.trim();
-  if (!raw) {
-    return 'custom';
-  }
-  const match = WORKFLOW_LIBRARY.find((entry) => entry.name.toLowerCase() === raw.toLowerCase());
-  return match ? match.name.toLowerCase() : raw;
-}
-
-function WorkflowKillButton({ onConfirm }: { onConfirm: () => void }) {
-  const [confirming, setConfirming] = useState(false);
-  if (confirming) {
-    return (
-      <span className="flex shrink-0 items-center gap-0.5 rounded-md border border-border bg-background/95 px-1 py-0.5 shadow-sm">
-        <span className="px-0.5 text-2xs text-muted-foreground">Discard?</span>
-        <button
-          type="button"
-          onClick={() => {
-            setConfirming(false);
-            onConfirm();
-          }}
-          title="confirm discard"
-          aria-label="confirm discard workflow"
-          className="rounded p-0.5 text-danger transition-colors hover:bg-danger/10"
-        >
-          <Check size={12} aria-hidden />
-        </button>
-        <button
-          type="button"
-          onClick={() => setConfirming(false)}
-          title="cancel"
-          aria-label="cancel discard workflow"
-          className="rounded p-0.5 text-muted-foreground transition-colors hover:bg-muted/50 hover:text-foreground"
-        >
-          <X size={12} aria-hidden />
-        </button>
-      </span>
-    );
-  }
-  return (
-    <button
-      type="button"
-      onClick={() => setConfirming(true)}
-      title="discard workflow"
-      aria-label="discard workflow"
-      className="shrink-0 rounded p-0.5 text-muted-foreground/60 transition-colors hover:bg-danger/10 hover:text-danger"
-    >
-      <Ban size={11} aria-hidden />
-    </button>
-  );
-}
-
-const pluralize = (count: number, singular: string) =>
-  `${count} ${singular}${count === 1 ? '' : 's'}`;
-
-function CollapsedSummary({ text }: { readonly text: string }) {
-  return <p className="pb-1 pl-2 text-2xs text-muted-foreground/60">{text}</p>;
-}
-
-type AgentsSectionProps = {
-  task: Session;
-};
-
-function AgentsSection({ task }: AgentsSectionProps) {
-  const isTaskActive = useAppStore((s) => s.currentSessionId === task.id);
-  const phaseRuns = useAppStore(
-    (s) => s.sessionPhaseRuns[task.id] ?? (EMPTY_ARRAY as ReadonlyArray<Agent>),
-  );
-  const telemetry = useAppStore(
-    (s) => s.sessionTelemetry[task.id] ?? (EMPTY_ARRAY as ReadonlyArray<TelemetryRecord>),
-  );
-  const messages = useAppStore((s) => s.messages[task.id] ?? EMPTY_ARRAY);
-  const agentRunHistory = useAppStore(
-    useShallow((s) => {
-      const out: Record<string, ReadonlyArray<ProviderRunId>> = {};
-      const runs = s.sessionPhaseRuns[task.id];
-      if (!runs) {
-        return out;
-      }
-      for (const run of runs) {
-        const history = s.agentRunHistory[run.id];
-        if (history) {
-          out[run.id] = history;
-        }
-      }
-      return out;
-    }),
-  );
-  const agentKindOverride = useAppStore(
-    useShallow((s) => {
-      const out: Record<string, AgentKind> = {};
-      const runs = s.sessionPhaseRuns[task.id];
-      if (!runs) {
-        return out;
-      }
-      for (const run of runs) {
-        const kind = s.agentKindOverride[run.id];
-        if (kind) {
-          out[run.id] = kind;
-        }
-      }
-      return out;
-    }),
-  );
-  const agentModelOverride = useAppStore(
-    useShallow((s) => {
-      const out: Record<string, string> = {};
-      const runs = s.sessionPhaseRuns[task.id];
-      if (!runs) {
-        return out;
-      }
-      for (const run of runs) {
-        const model = s.agentModelOverride[run.id];
-        if (model) {
-          out[run.id] = model;
-        }
-      }
-      return out;
-    }),
-  );
-  const selectedAgentId = useAppStore((s) => s.selectedAgentId[task.id] ?? null);
-  const worktreePath = useAppStore((s) => (s.sessionWorktrees[task.id] ?? [])[0] ?? null);
-  const prNumber = useAppStore((s) => s.sessionGithub[task.id]?.pr?.number ?? null);
-  const resolvedThreadIds = useAppStore(
-    useShallow(
-      (s) =>
-        new Set(
-          (s.sessionGithub[task.id]?.detail?.comments ?? [])
-            .filter((c) => c.resolved === true && c.threadId != null)
-            .map((c) => c.threadId as string),
-        ),
-    ),
-  );
-  const pendingThreadIds = useAppStore(
-    useShallow((s) => new Set((s.sessionPendingResolutions[task.id] ?? []).map((r) => r.threadId))),
-  );
-  const resolverState = useAppStore(
-    useShallow((s) => {
-      const out: Record<string, 'awaiting' | 'committed' | 'wontfix'> = {};
-      const runs = s.sessionPhaseRuns[task.id];
-      if (!runs) {
-        return out;
-      }
-      for (const run of runs) {
-        const st = s.resolverState[run.id];
-        if (st) {
-          out[run.id] = st;
-        }
-      }
-      return out;
-    }),
-  );
-  const selectAgent = useAppStore((s) => s.selectAgent);
-  const activateNextResolver = useAppStore((s) => s.activateNextResolver);
-  const spawnAgent = useAppStore((s) => s.spawnAgent);
-  const activateWorkflowAgent = useAppStore((s) => s.activateWorkflowAgent);
-  const renameAgent = useAppStore((s) => s.renameAgent);
-  const deleteAgent = useAppStore((s) => s.deleteAgent);
-  const phaseTemplates = useAppStore(
-    (s) => s.phaseTemplates[task.workspaceId] ?? (EMPTY_ARRAY as ReadonlyArray<Workflow>),
-  );
-  const sessionWorkflows = useAppStore(
-    (s) => s.sessionWorkflows[task.id] ?? (EMPTY_ARRAY as ReadonlyArray<Workflow>),
-  );
-  const attachedRuns = useMemo<ReadonlyArray<{ run: WorkflowRun; workflow: Workflow }>>(() => {
-    const byId = new Map<string, Workflow>();
-    for (const w of phaseTemplates) byId.set(w.id, w);
-    for (const w of sessionWorkflows) byId.set(w.id, w);
-    return [...task.workflowRuns]
-      .sort((a, b) => a.ordinal - b.ordinal)
-      .map((run) => {
-        const workflow = byId.get(run.workflowId);
-        return workflow ? { run, workflow } : null;
-      })
-      .filter((e): e is { run: WorkflowRun; workflow: Workflow } => e !== null);
-  }, [task.workflowRuns, phaseTemplates, sessionWorkflows]);
-  const discardWorkflow = useAppStore((s) => s.discardWorkflow);
-  const reorderSessionWorkflows = useAppStore((s) => s.reorderSessionWorkflows);
-  const setWorkflowRunAutoRun = useAppStore((s) => s.setWorkflowRunAutoRun);
-  const startWorkflowRun = useAppStore((s) => s.startWorkflowRun);
-  const workflowNameByRunId = useMemo(() => {
-    const map = new Map<string, string>();
-    for (const { run, workflow } of attachedRuns) {
-      map.set(run.id, workflowKindName(workflow));
-    }
-    return map;
-  }, [attachedRuns]);
-  const openQuestions = useSessionOpenQuestions(task.id);
-  const loading = useSessionLoading(task.id);
-  const summarizerBusy = useAppStore((s) => s.summarizerStatus[task.id]?.status === 'running');
-  const [spawnError, setSpawnError] = useState<string | null>(null);
-  const [editingId, setEditingId] = useState<AgentId | null>(null);
-  const [workflowExpand, setWorkflowExpand] = useState<ReadonlyMap<string, boolean>>(new Map());
-  const toggleWorkflowExpand = useCallback((id: string, isDiscarded: boolean) => {
-    setWorkflowExpand((prev) => {
-      const next = new Map(prev);
-      next.set(id, !(prev.get(id) ?? !isDiscarded));
-      return next;
-    });
-  }, []);
-  const [resolveExpanded, setResolveExpanded] = useState(true);
-  const [workflowExpanded, setWorkflowExpanded] = useState(true);
-  const [agentsExpanded, setAgentsExpanded] = useState(true);
-  const [clusterExpand, setClusterExpand] = useState<ReadonlyMap<string, boolean>>(new Map());
-  const toggleClusterExpand = useCallback((id: string) => {
-    setClusterExpand((prev) => {
-      const next = new Map(prev);
-      next.set(id, !(prev.get(id) ?? false));
-      return next;
-    });
-  }, []);
-
-  const sorted = useMemo(() => [...phaseRuns].sort((a, b) => a.ordinal - b.ordinal), [phaseRuns]);
-  const agentsByRunId = useMemo(() => {
-    const map = new Map<string, Agent[]>();
-    for (const r of sorted) {
-      if (r.stepId == null || r.workflowRunId == null) {
-        continue;
-      }
-      const bucket = map.get(r.workflowRunId) ?? [];
-      bucket.push(r);
-      map.set(r.workflowRunId, bucket);
-    }
-    return map;
-  }, [sorted]);
-  const childrenByParentId = useMemo(() => {
-    const map = new Map<string, Agent[]>();
-    for (const r of sorted) {
-      if (r.parentAgentId == null) {
-        continue;
-      }
-      const bucket = map.get(r.parentAgentId) ?? [];
-      bucket.push(r);
-      map.set(r.parentAgentId, bucket);
-    }
-    return map;
-  }, [sorted]);
-  const adHocAgents = useMemo(
-    () =>
-      sorted.filter(
-        (r) => r.parentAgentId == null && !(r.workflowRunId != null && r.stepId != null),
-      ),
-    [sorted],
-  );
-  const actionableStepIdByRunId = useMemo(() => {
-    const map = new Map<string, string | null>();
-    for (const { run, workflow } of attachedRuns) {
-      if (run.discardedAt) {
-        map.set(run.id, null);
-        continue;
-      }
-      const runAgents = agentsByRunId.get(run.id) ?? EMPTY_ARRAY;
-      map.set(run.id, pickNextWorkflowStep(workflow, runAgents)?.id ?? null);
-    }
-    return map;
-  }, [attachedRuns, agentsByRunId]);
-  const blockReasonByRunId = useMemo(() => {
-    const map = new Map<string, WorkflowBlockReason | null>();
-    for (const { run } of attachedRuns) {
-      const reason = workflowRunHasOpenQuestions(openQuestions, run.id)
-        ? 'questions'
-        : summarizerBusy
-          ? 'summarizer'
-          : null;
-      map.set(run.id, reason);
-    }
-    return map;
-  }, [attachedRuns, openQuestions, summarizerBusy]);
-
-  const onDiscardWorkflow = useCallback(
-    async (runId: WorkflowRunId) => {
-      try {
-        await discardWorkflow(task.id, runId);
-      } catch (err) {
-        setSpawnError(formatError(err));
-      }
-    },
-    [discardWorkflow, task.id],
-  );
-
-  const onReorderWorkflow = useCallback(
-    async (runId: WorkflowRunId, direction: 'up' | 'down') => {
-      const ids = [...task.workflowRuns].sort((a, b) => a.ordinal - b.ordinal).map((r) => r.id);
-      const idx = ids.indexOf(runId);
-      if (idx === -1) {
-        return;
-      }
-      const swap = direction === 'up' ? idx - 1 : idx + 1;
-      if (swap < 0 || swap >= ids.length) {
-        return;
-      }
-      [ids[idx], ids[swap]] = [ids[swap]!, ids[idx]!];
-      try {
-        await reorderSessionWorkflows(task.id, ids);
-      } catch (err) {
-        setSpawnError(formatError(err));
-      }
-    },
-    [reorderSessionWorkflows, task.id, task.workflowRuns],
-  );
-
-  const telemetryByRunId = useMemo(() => {
-    const map = new Map<string, TelemetryRecord>();
-    for (const rec of telemetry) {
-      const existing = map.get(rec.runId);
-      if (!existing || existing.recordedAt < rec.recordedAt) {
-        map.set(rec.runId, rec);
-      }
-    }
-    return map;
-  }, [telemetry]);
-
-  const turnsByAgentId = useMemo(() => {
-    const map = new Map<string, number>();
-    for (const m of messages) {
-      if (m.role !== 'user') {
-        continue;
-      }
-      map.set(m.agentId, (map.get(m.agentId) ?? 0) + 1);
-    }
-    return map;
-  }, [messages]);
-
-  const firstUserTextByAgentId = useMemo(() => {
-    const map = new Map<string, string>();
-    for (const m of messages) {
-      if (m.role !== 'user') {
-        continue;
-      }
-      if (map.has(m.agentId)) {
-        continue;
-      }
-      map.set(m.agentId, m.content);
-    }
-    return map;
-  }, [messages]);
-
-  const resolverAgents = useMemo(
-    () =>
-      sorted.filter(
-        (r) =>
-          r.parentAgentId == null &&
-          r.stepId == null &&
-          resolveAgentKind(
-            r.name,
-            firstUserTextByAgentId.get(r.id) ?? null,
-            agentKindOverride[r.id] ?? null,
-          ) === 'resolver',
-      ),
-    [sorted, firstUserTextByAgentId, agentKindOverride],
-  );
-  const resolverIds = useMemo(() => new Set(resolverAgents.map((r) => r.id)), [resolverAgents]);
-
-  const aggregatesByAgentId = useMemo(() => {
-    const map = new Map<
-      string,
-      { inputTokens: number; outputTokens: number; estimatedCostUsd: number; turns: number }
-    >();
-    const telemetryByRun = new Map<string, TelemetryRecord>();
-    for (const rec of telemetry) {
-      if (rec.kind !== 'turn') {
-        continue;
-      }
-      const existing = telemetryByRun.get(rec.runId);
-      if (!existing || existing.recordedAt < rec.recordedAt) {
-        telemetryByRun.set(rec.runId, rec);
-      }
-    }
-    for (const run of phaseRuns) {
-      const runIds = agentRunHistory[run.id] ?? (run.runId ? [run.runId] : []);
-      let inputTokens = 0;
-      let outputTokens = 0;
-      let estimatedCostUsd = 0;
-      let turns = 0;
-      for (const rid of runIds) {
-        const rec = telemetryByRun.get(rid);
-        if (!rec) {
-          continue;
-        }
-        inputTokens += rec.inputTokens;
-        outputTokens += rec.outputTokens;
-        estimatedCostUsd += rec.estimatedCostUsd;
-        turns += 1;
-      }
-      map.set(run.id, { inputTokens, outputTokens, estimatedCostUsd, turns });
-    }
-    const childIds = new Map<string, string[]>();
-    for (const run of phaseRuns) {
-      if (run.parentAgentId == null) {
-        continue;
-      }
-      const bucket = childIds.get(run.parentAgentId) ?? [];
-      bucket.push(run.id);
-      childIds.set(run.parentAgentId, bucket);
-    }
-    const rolled = new Set<string>();
-    const rollup = (id: string) => {
-      if (rolled.has(id)) {
-        return;
-      }
-      rolled.add(id);
-      const self = map.get(id);
-      if (!self) {
-        return;
-      }
-      for (const cid of childIds.get(id) ?? []) {
-        rollup(cid);
-        const child = map.get(cid);
-        if (!child) {
-          continue;
-        }
-        self.inputTokens += child.inputTokens;
-        self.outputTokens += child.outputTokens;
-        self.estimatedCostUsd += child.estimatedCostUsd;
-        self.turns += child.turns;
-      }
-    };
-    for (const run of phaseRuns) rollup(run.id);
-    return map;
-  }, [telemetry, phaseRuns, agentRunHistory]);
-
-  const latestTelemetryByAgentId = useMemo(
-    () => computeLatestTelemetryByAgentId(phaseRuns, agentRunHistory, telemetryByRunId),
-    [telemetryByRunId, phaseRuns, agentRunHistory],
-  );
-
-  const onPickAgent = (sid: AgentId) => {
-    if (sid === selectedAgentId) {
-      return;
-    }
-    void selectAgent(task.id, sid);
-  };
-
-  const onStartStepAgent = async (agent: Agent, model?: string) => {
-    setSpawnError(null);
-    try {
-      if (agent.status === 'pending') {
-        await activateWorkflowAgent(task.id, agent.id);
-        return;
-      }
-      await spawnAgent(task.id, {
-        ...(agent.stepId != null && { stepId: agent.stepId }),
-        ...(agent.workflowRunId != null && { workflowRunId: agent.workflowRunId }),
-        ...(model !== undefined && { model }),
-      });
-    } catch (err) {
-      setSpawnError(formatError(err));
-    }
-  };
-
-  const onRenameCommit = async (id: AgentId, name: string) => {
-    setEditingId(null);
-    try {
-      await renameAgent(task.id, id, name);
-    } catch (err) {
-      setSpawnError(formatError(err));
-    }
-  };
-
-  const onDeleteAgent = async (id: AgentId) => {
-    try {
-      await deleteAgent(task.id, id);
-    } catch (err) {
-      setSpawnError(formatError(err));
-    }
-  };
-
-  const renderAdHocRow = (run: Agent, index: number) => {
-    const kind = resolveAgentKind(
-      run.name,
-      firstUserTextByAgentId.get(run.id) ?? null,
-      agentKindOverride[run.id] ?? null,
-    );
-    const scoutChildren = childrenByParentId.get(run.id) ?? EMPTY_ARRAY;
-    return (
-      <Fragment key={run.id}>
-        <AgentRow
-          run={run}
-          kind={kind}
-          index={index}
-          telemetry={latestTelemetryByAgentId.get(run.id) ?? null}
-          aggregate={aggregatesByAgentId.get(run.id) ?? null}
-          turns={turnsByAgentId.get(run.id) ?? 0}
-          turnsLoading={run.id === selectedAgentId && loading.transcript}
-          isSelected={run.id === selectedAgentId}
-          isTaskActive={isTaskActive}
-          isEditing={editingId === run.id}
-          onClick={() => onPickAgent(run.id)}
-          onRenameStart={() => setEditingId(run.id)}
-          onRenameCommit={(name) => void onRenameCommit(run.id, name)}
-          onRenameCancel={() => setEditingId(null)}
-          onDelete={() => void onDeleteAgent(run.id)}
-        />
-        {scoutChildren.length > 0 && (
-          <li>
-            <ScoutSubtree
-              containerId={run.id}
-              depth={0}
-              childrenByParentId={childrenByParentId}
-              aggregatesByAgentId={aggregatesByAgentId}
-              selectedAgentId={selectedAgentId}
-              expandState={clusterExpand}
-              onToggle={toggleClusterExpand}
-              onSelect={onPickAgent}
-            />
-          </li>
-        )}
-      </Fragment>
-    );
-  };
-
-  const hasAnyWorkflow = attachedRuns.length > 0;
-  const renderWorkflowRow = (
-    { run, workflow }: { run: WorkflowRun; workflow: Workflow },
-    idx: number,
-  ) => {
-    const isDiscarded = run.discardedAt != null;
-    const expanded = workflowExpand.get(run.id) ?? !isDiscarded;
-    const wfAgents = agentsByRunId.get(run.id) ?? EMPTY_ARRAY;
-    const actionableStepId = actionableStepIdByRunId.get(run.id) ?? null;
-    const wfBlockReason = blockReasonByRunId.get(run.id) ?? null;
-    const canMoveUp = idx > 0;
-    const canMoveDown = idx < attachedRuns.length - 1;
-    const name = workflowKindName(workflow);
-    const total = workflow.steps.length;
-    const done = wfAgents.filter((a) => a.status === 'completed' || a.status === 'skipped').length;
-    const isCompleted = !isDiscarded && total > 0 && done >= total;
-    const hasStarted = wfAgents.length > 0;
-    const isQueuedManual = !isDiscarded && run.triggerMode === 'manual' && !hasStarted;
-    const isQueuedAfter = !isDiscarded && run.triggerMode === 'after_run' && !hasStarted;
-    const predecessorName = run.chainAfterId
-      ? (workflowNameByRunId.get(run.chainAfterId) ?? 'previous')
-      : 'previous';
-    return (
-      <div key={run.id} className={cn('flex flex-col', isDiscarded && 'opacity-70')}>
-        <div className="flex items-center gap-0.5">
-          <button
-            type="button"
-            onClick={() => toggleWorkflowExpand(run.id, isDiscarded)}
-            title={workflow.name || name}
-            aria-expanded={expanded}
-            aria-label={`${expanded ? 'collapse' : 'expand'} ${name} workflow`}
-            className="flex min-w-0 flex-1 items-center gap-1.5 rounded py-1 pl-1 pr-1.5 text-left transition-colors hover:bg-muted/50"
-          >
-            {expanded ? (
-              <ChevronDown size={12} aria-hidden className="shrink-0 text-muted-foreground/60" />
-            ) : (
-              <ChevronRight size={12} aria-hidden className="shrink-0 text-muted-foreground/60" />
-            )}
-            <span className="min-w-0 flex-1 truncate text-xs font-medium text-foreground">
-              {name}
-            </span>
-            {isDiscarded ? (
-              <span className="inline-flex shrink-0 items-center gap-1 rounded bg-muted px-1.5 py-0.5 text-[10px] font-medium uppercase tracking-wide text-muted-foreground">
-                <Ban size={10} aria-hidden /> discarded
-              </span>
-            ) : isCompleted ? (
-              <span className="inline-flex shrink-0 items-center gap-1 rounded bg-success/10 px-1.5 py-0.5 text-[10px] font-medium uppercase tracking-wide text-success">
-                <Check size={10} aria-hidden /> completed
-              </span>
-            ) : isQueuedManual ? (
-              <span className="inline-flex shrink-0 items-center gap-1 rounded bg-muted px-1.5 py-0.5 text-[10px] font-medium uppercase tracking-wide text-muted-foreground">
-                <Pause size={10} aria-hidden /> queued (manual)
-              </span>
-            ) : isQueuedAfter ? (
-              <span
-                className="inline-flex max-w-[10rem] shrink-0 items-center gap-1 truncate rounded bg-muted px-1.5 py-0.5 text-[10px] font-medium uppercase tracking-wide text-muted-foreground"
-                title={`after ${predecessorName}`}
-              >
-                <Link2 size={10} aria-hidden /> after {predecessorName}
-              </span>
-            ) : total > 0 ? (
-              <span className="shrink-0 font-mono text-[10px] text-muted-foreground/50">
-                {done}/{total}
-              </span>
-            ) : null}
-          </button>
-          {!isDiscarded && !isCompleted && (
-            <div className="flex shrink-0 items-center">
-              {isQueuedManual ? (
-                <button
-                  type="button"
-                  onClick={() => void startWorkflowRun(task.id, run.id)}
-                  title="start this workflow now"
-                  aria-label="start workflow now"
-                  className="rounded p-0.5 text-success transition-colors hover:bg-success/15"
-                >
-                  <Play size={11} aria-hidden />
-                </button>
-              ) : null}
-              <button
-                type="button"
-                onClick={() => void setWorkflowRunAutoRun(task.id, run.id, !run.autoRun)}
-                title={run.autoRun ? 'autorun on, click to pause' : 'autorun off, click to enable'}
-                aria-label={run.autoRun ? 'autorun on' : 'autorun off'}
-                aria-pressed={run.autoRun}
-                className={cn(
-                  'rounded p-0.5 transition-colors',
-                  run.autoRun
-                    ? 'text-danger hover:bg-danger/15'
-                    : 'text-muted-foreground/60 hover:bg-foreground/10 hover:text-foreground',
-                )}
-              >
-                {run.autoRun ? <Zap size={11} aria-hidden /> : <ZapOff size={11} aria-hidden />}
-              </button>
-              {attachedRuns.length > 1 && (
-                <>
-                  <button
-                    type="button"
-                    disabled={!canMoveUp}
-                    onClick={() => void onReorderWorkflow(run.id, 'up')}
-                    title="move workflow up"
-                    aria-label="move workflow up"
-                    className="rounded p-0.5 text-muted-foreground/60 transition-colors hover:bg-foreground/10 hover:text-foreground disabled:cursor-not-allowed disabled:opacity-30"
-                  >
-                    <ChevronUp size={11} aria-hidden />
-                  </button>
-                  <button
-                    type="button"
-                    disabled={!canMoveDown}
-                    onClick={() => void onReorderWorkflow(run.id, 'down')}
-                    title="move workflow down"
-                    aria-label="move workflow down"
-                    className="rounded p-0.5 text-muted-foreground/60 transition-colors hover:bg-foreground/10 hover:text-foreground disabled:cursor-not-allowed disabled:opacity-30"
-                  >
-                    <ChevronDown size={11} aria-hidden />
-                  </button>
-                </>
-              )}
-              <WorkflowKillButton onConfirm={() => void onDiscardWorkflow(run.id)} />
-            </div>
-          )}
-        </div>
-        {expanded ? (
-          wfAgents.length > 0 ? (
-            <div className="flex flex-col gap-1 pb-1 pl-3">
-              {wfAgents.map((run, index) => {
-                const isActionable = run.stepId === actionableStepId && run.status === 'pending';
-                const kind = agentKindOverride[run.id] ?? inferAgentKindFromName(run.name);
-                const stepModel =
-                  run.stepId != null
-                    ? workflow.steps.find((s) => s.id === run.stepId)?.modelOverride
-                    : undefined;
-                const resolvedModel =
-                  stepModel ??
-                  agentModelOverride[run.id] ??
-                  run.modelOverride ??
-                  AGENT_KIND_DEFAULTS[kind].model;
-                const clusterChildren = childrenByParentId.get(run.id) ?? EMPTY_ARRAY;
-                const clustersExpanded = clusterExpand.get(run.id) ?? false;
-                return (
-                  <Fragment key={run.id}>
-                    <WorkflowStepRow
-                      run={run}
-                      kind={kind}
-                      index={index}
-                      resolvedModel={resolvedModel}
-                      isActionable={isActionable}
-                      blockReason={isActionable ? wfBlockReason : null}
-                      isSelected={run.id === selectedAgentId}
-                      isTaskActive={isTaskActive}
-                      isEditing={editingId === run.id}
-                      telemetry={latestTelemetryByAgentId.get(run.id) ?? null}
-                      aggregate={aggregatesByAgentId.get(run.id) ?? null}
-                      turns={turnsByAgentId.get(run.id) ?? 0}
-                      turnsLoading={run.id === selectedAgentId && loading.transcript}
-                      onStart={() => void onStartStepAgent(run)}
-                      onSelect={() => onPickAgent(run.id)}
-                      onRenameStart={() => setEditingId(run.id)}
-                      onRenameCommit={(name) => void onRenameCommit(run.id, name)}
-                      onRenameCancel={() => setEditingId(null)}
-                    />
-                    {clusterChildren.length === 0 ? null : kind === 'scout' ? (
-                      <ScoutSubtree
-                        containerId={run.id}
-                        depth={0}
-                        childrenByParentId={childrenByParentId}
-                        aggregatesByAgentId={aggregatesByAgentId}
-                        selectedAgentId={selectedAgentId}
-                        expandState={clusterExpand}
-                        onToggle={toggleClusterExpand}
-                        onSelect={onPickAgent}
-                      />
-                    ) : (
-                      <div className="ml-3 flex flex-col gap-0.5 border-l border-border-soft/60 pl-2">
-                        <button
-                          type="button"
-                          onClick={() => toggleClusterExpand(run.id)}
-                          aria-expanded={clustersExpanded}
-                          aria-label={`${clustersExpanded ? 'collapse' : 'expand'} clusters for ${run.name}`}
-                          className="flex items-center gap-1 px-2 py-0.5 text-2xs uppercase tracking-wide text-muted-foreground/50 transition-colors hover:text-muted-foreground"
-                        >
-                          {clustersExpanded ? (
-                            <ChevronDown size={10} aria-hidden className="shrink-0" />
-                          ) : (
-                            <ChevronRight size={10} aria-hidden className="shrink-0" />
-                          )}
-                          <span className="min-w-0 truncate">
-                            clusters{' '}
-                            {clusterChildren.filter((c) => c.status === 'completed').length}/
-                            {clusterChildren.length}
-                          </span>
-                        </button>
-                        {clustersExpanded
-                          ? clusterChildren.map((child, ci) => (
-                              <ClusterChildRow
-                                key={child.id}
-                                child={child}
-                                index={ci}
-                                total={clusterChildren.length}
-                                costUsd={aggregatesByAgentId.get(child.id)?.estimatedCostUsd ?? 0}
-                                isSelected={child.id === selectedAgentId}
-                                onSelect={() => onPickAgent(child.id)}
-                              />
-                            ))
-                          : null}
-                      </div>
-                    )}
-                  </Fragment>
-                );
-              })}
-            </div>
-          ) : (
-            <p className="pb-1 pl-3 text-2xs text-muted-foreground/60">
-              no agents yet for this workflow.
-            </p>
-          )
-        ) : null}
-      </div>
-    );
-  };
-
-  return (
-    <section className="mt-2 flex flex-col px-2 pb-3">
-      <SectionHeader
-        className="pb-1.5"
-        icon={<Layers size={11} aria-hidden className="text-primary" />}
-        label="Workflow"
-        action={
-          <SectionToggle
-            expanded={workflowExpanded}
-            label="workflow"
-            onToggle={() => setWorkflowExpanded((v) => !v)}
-          />
-        }
-      />
-      {workflowExpanded ? (
-        !hasAnyWorkflow ? (
-          <WorkflowStartButton sessionId={task.id} variant="empty" />
-        ) : (
-          <>
-            <div className="flex flex-col gap-0.5">{attachedRuns.map(renderWorkflowRow)}</div>
-            <WorkflowStartButton sessionId={task.id} variant="attach" />
-          </>
-        )
-      ) : (
-        <CollapsedSummary
-          text={hasAnyWorkflow ? pluralize(attachedRuns.length, 'workflow') : 'No workflows yet'}
-        />
-      )}
-
-      <SectionHeader
-        className="mt-6 pb-1.5"
-        icon={<DogMascot size={14} className="shrink-0 text-success" />}
-        label="Agents"
-        action={
-          <SectionToggle
-            expanded={agentsExpanded}
-            label="agents"
-            onToggle={() => setAgentsExpanded((v) => !v)}
-          />
-        }
-      />
-      {agentsExpanded ? (
-        <>
-          {hasAnyWorkflow ? (
-            adHocAgents.some((r) => !resolverIds.has(r.id)) ? (
-              <ul className="flex flex-col gap-1 pl-2">
-                {adHocAgents.filter((r) => !resolverIds.has(r.id)).map(renderAdHocRow)}
-              </ul>
-            ) : null
-          ) : sorted.length === 0 ? (
-            loading.agents ? (
-              <ul role="status" aria-label="loading agents" className="flex flex-col gap-1 pl-2">
-                {[0, 1].map((i) => (
-                  <li key={i} className="flex items-center gap-2 rounded px-2 py-1.5">
-                    <span className="h-3 w-3 animate-pulse rounded-full bg-muted" />
-                    <span className="h-3 flex-1 animate-pulse rounded bg-muted" />
-                  </li>
-                ))}
-              </ul>
-            ) : resolverAgents.length === 0 ? (
-              <p className="px-2 py-2 text-xs text-muted-foreground/70">
-                No agents yet. Spawn one below.
-              </p>
-            ) : null
-          ) : (
-            <ul className="flex flex-col gap-1 pl-2">
-              {sorted.filter((r) => !resolverIds.has(r.id)).map(renderAdHocRow)}
-            </ul>
-          )}
-          <div className="flex flex-col gap-1">
-            <PlanReadySuggestion task={task} />
-            <SpawnAgentControl sessionId={task.id} />
-          </div>
-          {spawnError ? <p className="mt-1 px-2 text-2xs text-danger">{spawnError}</p> : null}
-        </>
-      ) : (
-        <CollapsedSummary
-          text={(() => {
-            const count = sorted.filter((r) => !resolverIds.has(r.id)).length;
-            return count === 0 ? 'No agents yet' : pluralize(count, 'agent');
-          })()}
-        />
-      )}
-
-      <ScriptsSection
-        sessionId={task.id}
-        workspaceId={task.workspaceId}
-        worktreePath={worktreePath}
-      />
-
-      {resolverAgents.length > 0 && (
-        <>
-          <SectionHeader
-            className="mt-6 pb-1.5"
-            icon={<CheckCheck size={11} aria-hidden className="text-lime-500" />}
-            label="Resolve"
-          />
-          <ResolveCluster
-            agents={resolverAgents}
-            sessionId={task.id}
-            isTaskActive={isTaskActive}
-            prNumber={prNumber}
-            resolvedThreadIds={resolvedThreadIds}
-            pendingThreadIds={pendingThreadIds}
-            resolverState={resolverState}
-            selectedAgentId={selectedAgentId}
-            expanded={resolveExpanded}
-            onToggle={() => setResolveExpanded((v) => !v)}
-            onSelect={onPickAgent}
-            onForceNext={() => void activateNextResolver(task.id)}
-          />
-        </>
-      )}
-    </section>
-  );
-}
-
-type SpawnAgentControlProps = {
-  sessionId: SessionId;
-};
-
-type PopoverAnchor = {
-  readonly left: number;
-  readonly top: number | null;
-  readonly bottom: number | null;
-  readonly direction: 'up' | 'down';
-};
-
-function SpawnAgentControl({ sessionId }: SpawnAgentControlProps) {
-  const [open, setOpen] = useState(false);
-  const [anchor, setAnchor] = useState<PopoverAnchor | null>(null);
-  const triggerRef = useRef<HTMLButtonElement>(null);
-  const menuRef = useRef<HTMLDivElement>(null);
-  const spawnAgent = useAppStore((s) => s.spawnAgent);
-
-  const computeAnchor = useCallback((): PopoverAnchor | null => {
-    const rect = triggerRef.current?.getBoundingClientRect();
-    if (!rect) {
-      return null;
-    }
-    const spaceAbove = rect.top;
-    const spaceBelow = window.innerHeight - rect.bottom;
-    const direction: 'up' | 'down' = spaceBelow > spaceAbove ? 'down' : 'up';
-    const left = rect.right + 4;
-    if (direction === 'down') {
-      return { left, top: rect.top, bottom: null, direction };
-    }
-    return { left, top: null, bottom: window.innerHeight - rect.bottom, direction };
-  }, []);
-
-  useEffect(() => {
-    if (!open) {
-      return;
-    }
-    const onDocClick = (e: MouseEvent) => {
-      const target = e.target as Node;
-      if (triggerRef.current?.contains(target)) {
-        return;
-      }
-      if (menuRef.current?.contains(target)) {
-        return;
-      }
-      setOpen(false);
-    };
-    const onReanchor = () => {
-      const next = computeAnchor();
-      if (next) {
-        setAnchor(next);
-      } else {
-        setOpen(false);
-      }
-    };
-    window.addEventListener('mousedown', onDocClick);
-    window.addEventListener('resize', onReanchor);
-    window.addEventListener('scroll', onReanchor, true);
-    return () => {
-      window.removeEventListener('mousedown', onDocClick);
-      window.removeEventListener('resize', onReanchor);
-      window.removeEventListener('scroll', onReanchor, true);
-    };
-  }, [open, computeAnchor]);
-
-  const onToggle = () => {
-    if (!open) {
-      const next = computeAnchor();
-      if (next) {
-        setAnchor(next);
-      }
-    }
-    setOpen((v) => !v);
-  };
-
-  const menu =
-    open && anchor
-      ? createPortal(
-          <Popover
-            innerRef={menuRef}
-            role="menu"
-            style={{
-              position: 'fixed',
-              left: anchor.left,
-              ...(anchor.top !== null ? { top: anchor.top } : {}),
-              ...(anchor.bottom !== null ? { bottom: anchor.bottom } : {}),
-            }}
-            className="z-50 w-80 max-h-72 overflow-y-auto py-1"
-          >
-            <div className="px-2.5 pb-1 pt-1.5 text-2xs uppercase tracking-wide text-muted-foreground/70">
-              by role
-            </div>
-            {[...AGENT_KIND_ORDER]
-              .filter((kind) => AGENT_KIND_DEFAULTS[kind].visible !== false)
-              .sort((a, b) => AGENT_KIND_META[a].label.localeCompare(AGENT_KIND_META[b].label))
-              .map((kind) => {
-                const meta = AGENT_KIND_META[kind];
-                const palette = AGENT_KIND_PALETTE[kind];
-                return (
-                  <button
-                    key={kind}
-                    type="button"
-                    role="menuitem"
-                    onClick={() => {
-                      setOpen(false);
-                      void spawnAgent(sessionId, { kindOverride: kind });
-                    }}
-                    className="flex w-full items-center gap-2 px-2.5 py-1.5 text-left transition-colors hover:bg-muted"
-                  >
-                    <span className={cn('size-2 shrink-0 rounded-full', palette.bg)} aria-hidden />
-                    <span className="font-medium text-foreground">{meta.label}</span>
-                    <span className="truncate text-2xs text-muted-foreground">{meta.hint}</span>
-                  </button>
-                );
-              })}
-          </Popover>,
-          document.body,
-        )
-      : null;
-
-  return (
-    <div className="relative mt-1">
-      <button
-        ref={triggerRef}
-        type="button"
-        onClick={onToggle}
-        className="flex w-full items-center gap-2 rounded border border-dashed border-border-soft px-2 py-1.5 text-left text-xs text-muted-foreground transition-colors hover:border-border hover:bg-muted/50 hover:text-foreground"
-        aria-haspopup="menu"
-        aria-expanded={open}
-      >
-        <Plus size={13} aria-hidden />
-        Create agent
-      </button>
-      {menu}
-    </div>
-  );
-}
-
-type ClusterChildRowProps = {
-  readonly child: Agent;
-  readonly index: number;
-  readonly total: number;
-  readonly costUsd: number;
-  readonly isSelected: boolean;
-  readonly onSelect: () => void;
-};
-
-function ClusterChildRow({
-  child,
-  index,
-  total,
-  costUsd,
-  isSelected,
-  onSelect,
-}: ClusterChildRowProps) {
-  const icon =
-    child.status === 'running' ? (
-      <Loader2 size={10} className="animate-spin text-info" aria-hidden />
-    ) : child.status === 'completed' ? (
-      <span className="flex size-3 items-center justify-center rounded-full bg-success/15">
-        <Check size={8} className="text-success" aria-hidden />
-      </span>
-    ) : child.status === 'failed' ? (
-      <span className="size-1.5 rounded-full bg-danger" aria-hidden />
-    ) : (
-      <Clock size={10} className="text-muted-foreground/60" aria-hidden />
-    );
-  return (
-    <button
-      type="button"
-      onClick={onSelect}
-      className={cn(
-        'flex w-full items-center gap-2 rounded px-2 py-1 text-2xs font-medium transition-colors',
-        isSelected
-          ? 'bg-elevated text-foreground'
-          : 'text-foreground/70 hover:bg-muted/60 hover:text-foreground',
-      )}
-    >
-      <span className="tabular-nums text-muted-foreground/50">
-        {index + 1}/{total}
-      </span>
-      {icon}
-      <span className="min-w-0 flex-1 truncate text-left">{child.name}</span>
-      {costUsd > 0 && (
-        <span
-          className="shrink-0 tabular-nums text-muted-foreground/60"
-          title={`$${costUsd.toFixed(4)}`}
-        >
-          ${costUsd.toFixed(2)}
-        </span>
-      )}
-    </button>
-  );
-}
-
-type ScoutSubtreeProps = {
-  readonly containerId: AgentId;
-  readonly depth: number;
-  readonly childrenByParentId: ReadonlyMap<string, Agent[]>;
-  readonly aggregatesByAgentId: ReadonlyMap<string, AgentAggregate>;
-  readonly selectedAgentId: AgentId | null;
-  readonly expandState: ReadonlyMap<string, boolean>;
-  readonly onToggle: (id: string) => void;
-  readonly onSelect: (id: AgentId) => void;
-};
-
-function ScoutSubtree({
-  containerId,
-  depth,
-  childrenByParentId,
-  aggregatesByAgentId,
-  selectedAgentId,
-  expandState,
-  onToggle,
-  onSelect,
-}: ScoutSubtreeProps) {
-  const children = childrenByParentId.get(containerId) ?? EMPTY_ARRAY;
-  if (children.length === 0 || depth > 4) {
-    return null;
-  }
-  const expanded = expandState.get(containerId) ?? false;
-  const doneCount = children.filter(
-    (c) => c.status === 'completed' || c.status === 'skipped',
-  ).length;
-  return (
-    <div className="ml-3 flex flex-col gap-0.5 border-l border-border-soft/60 pl-2">
-      <button
-        type="button"
-        onClick={() => onToggle(containerId)}
-        aria-expanded={expanded}
-        aria-label={`${expanded ? 'collapse' : 'expand'} scouts`}
-        className="flex items-center gap-1 px-2 py-0.5 text-2xs uppercase tracking-wide text-sky-400/70 transition-colors hover:text-sky-400"
-      >
-        {expanded ? (
-          <ChevronDown size={10} aria-hidden className="shrink-0" />
-        ) : (
-          <ChevronRight size={10} aria-hidden className="shrink-0" />
-        )}
-        scouts {doneCount}/{children.length}
-      </button>
-      {expanded
-        ? children.map((child, ci) => (
-            <Fragment key={child.id}>
-              <ClusterChildRow
-                child={child}
-                index={ci}
-                total={children.length}
-                costUsd={aggregatesByAgentId.get(child.id)?.estimatedCostUsd ?? 0}
-                isSelected={child.id === selectedAgentId}
-                onSelect={() => onSelect(child.id)}
-              />
-              <ScoutSubtree
-                containerId={child.id}
-                depth={depth + 1}
-                childrenByParentId={childrenByParentId}
-                aggregatesByAgentId={aggregatesByAgentId}
-                selectedAgentId={selectedAgentId}
-                expandState={expandState}
-                onToggle={onToggle}
-                onSelect={onSelect}
-              />
-            </Fragment>
-          ))
-        : null}
-    </div>
-  );
-}
-
-type ResolverState = 'awaiting' | 'committed' | 'wontfix';
-type ResolverStatus =
-  | 'running'
-  | 'failed'
-  | 'pending'
-  | 'resolved'
-  | 'committed'
-  | 'wontfix'
-  | 'awaiting'
-  | 'done';
-
-function resolverStatus(
-  agent: Agent,
-  resolvedThreadIds: ReadonlySet<string>,
-  pendingThreadIds: ReadonlySet<string>,
-  state: ResolverState | undefined,
-): ResolverStatus {
-  if (agent.status === 'running') {
-    return 'running';
-  }
-  if (agent.status === 'failed') {
-    return 'failed';
-  }
-  if (agent.status === 'pending') {
-    return 'pending';
-  }
-  const tid = agent.sourceThreadId;
-  if (tid != null && resolvedThreadIds.has(tid)) {
-    return 'resolved';
-  }
-  if (state === 'committed' || (tid != null && pendingThreadIds.has(tid))) {
-    return 'committed';
-  }
-  if (state === 'wontfix') {
-    return 'wontfix';
-  }
-  if (state === 'awaiting') {
-    return 'awaiting';
-  }
-  return 'done';
-}
-
-type ResolveClusterProps = {
-  readonly agents: ReadonlyArray<Agent>;
-  readonly sessionId: SessionId;
-  readonly isTaskActive: boolean;
-  readonly prNumber: number | null;
-  readonly resolvedThreadIds: ReadonlySet<string>;
-  readonly pendingThreadIds: ReadonlySet<string>;
-  readonly resolverState: Readonly<Record<string, ResolverState>>;
-  readonly selectedAgentId: AgentId | null;
-  readonly expanded: boolean;
-  readonly onToggle: () => void;
-  readonly onSelect: (id: AgentId) => void;
-  readonly onForceNext: () => void;
-};
-
-function ResolveCluster({
-  agents,
-  sessionId,
-  isTaskActive,
-  prNumber,
-  resolvedThreadIds,
-  pendingThreadIds,
-  resolverState,
-  selectedAgentId,
-  expanded,
-  onToggle,
-  onSelect,
-  onForceNext,
-}: ResolveClusterProps) {
-  const statusOf = (a: Agent): ResolverStatus =>
-    resolverStatus(a, resolvedThreadIds, pendingThreadIds, resolverState[a.id]);
-  const resolvedCount = agents.filter((a) => statusOf(a) === 'resolved').length;
-  const anyRunning = agents.some((a) => a.status === 'running');
-  const queuedCount = agents.filter((a) => a.status === 'pending').length;
-  const stalled = !anyRunning && queuedCount > 0;
-  const jump = (agent: Agent) => {
-    if (agent.sourceThreadId != null && prNumber != null) {
-      window.dispatchEvent(
-        new CustomEvent('goodboy:open-github-session', {
-          detail: { sessionId, prNumber, threadId: agent.sourceThreadId },
-        }),
-      );
-    } else if (agent.sourceCommentUrl != null) {
-      void openUrl(agent.sourceCommentUrl);
-    }
-  };
-  return (
-    <div className="flex flex-col gap-0.5 pl-2">
-      <div className="flex items-center gap-1 pr-1">
-        <button
-          type="button"
-          onClick={onToggle}
-          aria-expanded={expanded}
-          aria-label={`${expanded ? 'collapse' : 'expand'} resolve cluster`}
-          className="flex min-w-0 flex-1 items-center gap-1 px-2 py-0.5 text-2xs uppercase tracking-wide text-lime-500/80 transition-colors hover:text-lime-500"
-        >
-          {expanded ? (
-            <ChevronDown size={10} aria-hidden className="shrink-0" />
-          ) : (
-            <ChevronRight size={10} aria-hidden className="shrink-0" />
-          )}
-          {resolvedCount}/{agents.length} resolved
-        </button>
-        {stalled ? (
-          <button
-            type="button"
-            onClick={onForceNext}
-            title="the current resolver has not committed or explained yet; run the next queued one anyway"
-            className="inline-flex shrink-0 items-center gap-1 rounded border border-warning/40 bg-warning/10 px-1.5 py-0.5 text-[10px] font-medium text-warning transition-colors hover:bg-warning/20"
-          >
-            <Play size={9} aria-hidden />
-            Run next ({queuedCount})
-          </button>
-        ) : null}
-      </div>
-      {expanded
-        ? agents.map((agent, i) => (
-            <ResolveClusterRow
-              key={agent.id}
-              agent={agent}
-              index={i}
-              total={agents.length}
-              status={statusOf(agent)}
-              isSelected={agent.id === selectedAgentId}
-              isTaskActive={isTaskActive}
-              canJump={agent.sourceThreadId != null || agent.sourceCommentUrl != null}
-              onSelect={() => onSelect(agent.id)}
-              onJump={() => jump(agent)}
-            />
-          ))
-        : null}
-    </div>
-  );
-}
-
-type ResolveClusterRowProps = {
-  readonly agent: Agent;
-  readonly index: number;
-  readonly total: number;
-  readonly status: ResolverStatus;
-  readonly isSelected: boolean;
-  readonly isTaskActive: boolean;
-  readonly canJump: boolean;
-  readonly onSelect: () => void;
-  readonly onJump: () => void;
-};
-
-function ResolveClusterRow({
-  agent,
-  index,
-  total,
-  status,
-  isSelected,
-  isTaskActive,
-  canJump,
-  onSelect,
-  onJump,
-}: ResolveClusterRowProps) {
-  const hasUnread = agentHasUnread(agent, isSelected && isTaskActive);
-  const icon =
-    status === 'running' ? (
-      <Loader2 size={10} className="animate-spin text-info" aria-hidden />
-    ) : status === 'failed' ? (
-      <span className="size-1.5 rounded-full bg-danger" aria-hidden />
-    ) : status === 'pending' ? (
-      <Clock size={10} className="text-muted-foreground/60" aria-hidden />
-    ) : status === 'resolved' ? (
-      <CheckCheck size={10} className="text-success" aria-hidden />
-    ) : status === 'committed' ? (
-      <GitCommit size={10} className="text-warning" aria-hidden />
-    ) : status === 'wontfix' ? (
-      <Ban size={10} className="text-muted-foreground/70" aria-hidden />
-    ) : status === 'awaiting' ? (
-      <AlertTriangle size={10} className="text-warning" aria-hidden />
-    ) : (
-      <Check size={10} className="text-muted-foreground/70" aria-hidden />
-    );
-  const statusLabel =
-    status === 'resolved'
-      ? 'resolved on GitHub'
-      : status === 'committed'
-        ? 'committed locally, pending push'
-        : status === 'wontfix'
-          ? 'explained, pending resolve'
-          : status === 'awaiting'
-            ? 'needs you: no commit yet'
-            : status === 'running'
-              ? 'working'
-              : status === 'pending'
-                ? 'queued'
-                : status === 'failed'
-                  ? 'failed'
-                  : 'done locally';
-  return (
-    <div
-      className={cn(
-        'flex w-full items-center gap-2 rounded border px-2 py-1 text-2xs font-medium transition-colors',
-        isSelected
-          ? 'bg-elevated text-foreground border-border'
-          : 'text-foreground/70 hover:bg-muted/60',
-        status === 'running'
-          ? 'border-info/60'
-          : status === 'awaiting' || hasUnread
-            ? 'border-warning/70'
-            : 'border-transparent',
-      )}
-    >
-      <span className="tabular-nums text-muted-foreground/50">
-        {index + 1}/{total}
-      </span>
-      <span title={statusLabel}>{icon}</span>
-      <button
-        type="button"
-        onClick={onSelect}
-        className="min-w-0 flex-1 truncate text-left hover:text-foreground"
-      >
-        {agent.name}
-      </button>
-      {canJump ? (
-        <button
-          type="button"
-          onClick={onJump}
-          title="go to the review comment"
-          aria-label="go to the review comment"
-          className="shrink-0 rounded p-0.5 text-muted-foreground/60 transition-colors hover:bg-foreground/10 hover:text-foreground"
-        >
-          <MessageSquareReply size={11} aria-hidden />
-        </button>
-      ) : null}
-    </div>
-  );
-}
-
-type WorkflowBlockReason = 'questions' | 'summarizer';
-
-type WorkflowStepRowProps = {
-  readonly run: Agent;
-  readonly kind: AgentKind;
-  readonly index: number;
-  readonly resolvedModel: string;
-  readonly isActionable: boolean;
-  readonly blockReason: WorkflowBlockReason | null;
-  readonly isSelected: boolean;
-  readonly isTaskActive: boolean;
-  readonly isEditing: boolean;
-  readonly telemetry: TelemetryRecord | null;
-  readonly aggregate: AgentAggregate | null;
-  readonly turns: number;
-  readonly turnsLoading: boolean;
-  readonly onStart: () => void;
-  readonly onSelect: () => void;
-  readonly onRenameStart: () => void;
-  readonly onRenameCommit: (name: string) => void;
-  readonly onRenameCancel: () => void;
-};
-
-function WorkflowStepRow({
-  run,
-  kind,
-  index,
-  resolvedModel,
-  isActionable,
-  blockReason,
-  isSelected,
-  isTaskActive,
-  isEditing,
-  telemetry,
-  aggregate,
-  turns,
-  turnsLoading,
-  onStart,
-  onSelect,
-  onRenameStart,
-  onRenameCommit,
-  onRenameCancel,
-}: WorkflowStepRowProps) {
-  const isBlocked = blockReason !== null;
-  const isPendingFuture = run.status === 'pending' && !isActionable;
-  const modelLabel = resolvedModel.split('-').slice(1, 3).join('-');
-  const isStartable = isActionable && !isBlocked;
-  const stepCost = aggregate?.estimatedCostUsd ?? telemetry?.estimatedCostUsd ?? 0;
-
-  const [draft, setDraft] = useState(run.name);
-  const [pendingConfirm, setPendingConfirm] = useState(false);
-  useEffect(() => {
-    if (isEditing) {
-      setDraft(run.name);
-    }
-  }, [isEditing, run.name]);
-  useEffect(() => {
-    if (!isBlocked) {
-      setPendingConfirm(false);
-    }
-  }, [isBlocked]);
-
-  const handleRowClick = () => {
-    if (isPendingFuture) {
-      return;
-    }
-    if (isStartable) {
-      onStart();
-    } else if (isActionable && isBlocked) {
-      setPendingConfirm(true);
-    } else {
-      onSelect();
-    }
-  };
-
-  const ROW_BASE = 'group flex w-full flex-col rounded border px-2 py-1.5 text-xs font-medium';
-  const isRunning = run.status === 'running';
-  const hasUnread = agentHasUnread(run, isSelected && isTaskActive);
-  const containerClass = isRunning
-    ? cn(
-        `${ROW_BASE} border-info/60 transition-colors cursor-pointer`,
-        isSelected ? 'bg-elevated text-foreground' : 'bg-muted/40 text-foreground/80',
-      )
-    : isStartable
-      ? `${ROW_BASE} border-primary/40 bg-primary/10 text-primary shadow-sm transition-colors hover:border-primary hover:bg-primary/20 cursor-pointer`
-      : isActionable && isBlocked
-        ? `${ROW_BASE} border-warning/70 bg-warning/10 text-foreground transition-colors hover:bg-warning/15 cursor-pointer`
-        : hasUnread
-          ? `${ROW_BASE} border-warning/70 bg-muted/40 text-foreground/80 transition-colors hover:border-warning hover:bg-muted/60 cursor-pointer`
-          : isPendingFuture
-            ? `${ROW_BASE} border-transparent text-muted-foreground/40`
-            : cn(
-                `${ROW_BASE} transition-colors cursor-pointer`,
-                isSelected
-                  ? 'border-border bg-elevated text-foreground'
-                  : 'border-border-soft/50 bg-muted/40 text-foreground/80 hover:border-border hover:bg-muted/60',
-              );
-
-  const renderStatusIcon = () => {
-    if (isStartable) {
-      return (
-        <span className="relative inline-flex size-3.5">
-          <span
-            className="absolute inset-0 animate-ping rounded-full bg-primary/30 opacity-75"
-            aria-hidden
-          />
-          <span className="relative flex size-3.5 items-center justify-center rounded-full bg-primary/15">
-            <Play size={9} className="text-primary" aria-hidden fill="currentColor" />
-          </span>
-        </span>
-      );
-    }
-    if (isActionable && isBlocked) {
-      return <AlertTriangle size={12} className="text-warning" aria-hidden />;
-    }
-    if (run.status === 'running') {
-      return <Loader2 size={11} className="animate-spin text-info" aria-hidden />;
-    }
-    if (run.status === 'completed') {
-      return (
-        <span className="flex size-3.5 items-center justify-center rounded-full bg-success/15">
-          <Check size={9} className="text-success" aria-hidden />
-        </span>
-      );
-    }
-    if (run.status === 'failed') {
-      return <span className="size-1.5 rounded-full bg-danger" aria-hidden />;
-    }
-    return <Clock size={11} className="text-muted-foreground/70" aria-hidden />;
-  };
-
-  const renderActionIndicator = () => {
-    if (run.status === 'running') {
-      return null;
-    }
-    if (isActionable && isBlocked) {
-      return <ArrowRight size={13} aria-hidden className="text-warning" />;
-    }
-    return null;
-  };
-
-  const stableTitle =
-    isActionable && isBlocked
-      ? blockReason === 'summarizer'
-        ? 'next workflow step. waiting for the summarizer to finish (click to force)'
-        : 'next workflow step. gated by open questions (click to force)'
-      : isPendingFuture
-        ? 'waiting for previous steps'
-        : `agent ${run.ordinal + 1}: ${run.status}`;
-
-  return (
-    <div className="flex flex-col gap-1">
-      <div
-        role={isPendingFuture ? undefined : 'button'}
-        tabIndex={isEditing || isPendingFuture ? -1 : 0}
-        aria-pressed={isPendingFuture ? undefined : isSelected}
-        title={stableTitle}
-        onClick={isEditing || isPendingFuture ? undefined : handleRowClick}
-        onDoubleClick={isEditing || isPendingFuture ? undefined : onRenameStart}
-        onKeyDown={(e) => {
-          if (isEditing) {
-            return;
-          }
-          if (e.key === 'Enter' || e.key === ' ') {
-            e.preventDefault();
-            handleRowClick();
-          }
-        }}
-        className={containerClass}
-      >
-        <div className="flex w-full items-center gap-1.5">
-          <span
-            aria-hidden
-            className={cn(
-              'w-4 shrink-0 text-right text-2xs tabular-nums',
-              isPendingFuture ? 'text-muted-foreground/40' : 'text-muted-foreground/60',
-            )}
-          >
-            {index + 1}.
-          </span>
-          <span className="flex size-3.5 shrink-0 items-center justify-center">
-            {renderStatusIcon()}
-          </span>
-          <AgentKindChip kind={kind} muted={isPendingFuture} />
-          {isEditing ? (
-            <input
-              autoFocus
-              value={draft}
-              onChange={(e) => setDraft(e.target.value)}
-              onClick={(e) => e.stopPropagation()}
-              onDoubleClick={(e) => e.stopPropagation()}
-              onKeyDown={(e) => {
-                e.stopPropagation();
-                if (e.key === 'Enter') {
-                  e.preventDefault();
-                  onRenameCommit(draft);
-                } else if (e.key === 'Escape') {
-                  e.preventDefault();
-                  onRenameCancel();
-                }
-              }}
-              onBlur={() => onRenameCommit(draft)}
-              className="min-w-0 flex-1 rounded bg-background px-1.5 py-0.5 text-xs font-semibold text-foreground outline-none ring-1 ring-primary"
-              aria-label="rename agent"
-            />
-          ) : (
-            <span className="min-w-0 flex-1 truncate text-left font-semibold">{run.name}</span>
-          )}
-          {renderActionIndicator()}
-        </div>
-        <div className="flex w-full items-center justify-between gap-2 pl-[22px] pt-0.5">
-          <span className="flex min-w-0 items-center gap-1.5">
-            <ProviderGlyph
-              provider={telemetry?.provider ?? getModelProvider(resolvedModel)}
-              muted={isPendingFuture}
-            />
-            <span
-              className={cn(
-                'min-w-0 truncate text-[10px] font-normal tabular-nums',
-                isPendingFuture ? 'text-muted-foreground/60' : 'opacity-60',
-              )}
-              title={`model: ${resolvedModel}`}
-            >
-              {modelLabel}
-            </span>
-          </span>
-          {stepCost > 0 && (
-            <span
-              className="shrink-0 font-mono text-[10px] tabular-nums text-muted-foreground/60"
-              title={`estimated cost: ${formatCost(stepCost)}`}
-            >
-              {formatCost(stepCost)}
-            </span>
-          )}
-        </div>
-        <div
-          className={cn(
-            'grid w-full transition-[grid-template-rows] duration-200 ease-out',
-            isSelected && !isPendingFuture ? 'grid-rows-[1fr]' : 'grid-rows-[0fr]',
-          )}
-        >
-          <div className="overflow-hidden">
-            <div className="flex flex-col gap-1 pt-1">
-              <AgentMetricsBlock
-                run={run}
-                telemetry={telemetry}
-                aggregate={aggregate}
-                turns={turns}
-                turnsLoading={turnsLoading}
-                variant="workflow"
-              />
-              <ContextWindowBar telemetry={telemetry} aggregate={aggregate} />
-            </div>
-          </div>
-        </div>
-      </div>
-      {pendingConfirm ? (
-        <div className="flex items-center gap-2 rounded-md bg-warning/5 px-2.5 py-1.5 text-[11px]">
-          <AlertTriangle size={12} aria-hidden className="shrink-0 text-warning" />
-          <span className="min-w-0 flex-1 truncate text-foreground">
-            {blockReason === 'summarizer' ? 'Summarizer still running.' : 'Open questions first.'}
-          </span>
-          <button
-            type="button"
-            onClick={(e) => {
-              e.stopPropagation();
-              setPendingConfirm(false);
-            }}
-            className="shrink-0 rounded px-1.5 py-0.5 text-[10px] font-medium text-muted-foreground transition-colors hover:bg-muted hover:text-foreground"
-          >
-            {blockReason === 'summarizer' ? 'Wait' : 'Resolve first'}
-          </button>
-          <button
-            type="button"
-            onClick={(e) => {
-              e.stopPropagation();
-              setPendingConfirm(false);
-              onStart();
-            }}
-            className="shrink-0 rounded px-1.5 py-0.5 text-[10px] font-semibold text-warning transition-colors hover:bg-warning/10"
-          >
-            Force start
-          </button>
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-type AgentRowProps = {
-  readonly run: Agent;
-  readonly kind: AgentKind;
-  readonly index: number;
-  readonly telemetry: TelemetryRecord | null;
-  readonly aggregate: AgentAggregate | null;
-  readonly turns: number;
-  readonly turnsLoading: boolean;
-  readonly isSelected: boolean;
-  readonly isTaskActive: boolean;
-  readonly isEditing: boolean;
-  readonly onClick: () => void;
-  readonly onRenameStart: () => void;
-  readonly onRenameCommit: (name: string) => void;
-  readonly onRenameCancel: () => void;
-  readonly onDelete: () => void;
-};
-
-function AgentRow({
-  run,
-  kind,
-  index,
-  telemetry,
-  aggregate,
-  turns,
-  turnsLoading,
-  isSelected,
-  isTaskActive,
-  isEditing,
-  onClick,
-  onRenameStart,
-  onRenameCommit,
-  onRenameCancel,
-  onDelete,
-}: AgentRowProps) {
-  const total = telemetry ? telemetry.inputTokens + telemetry.outputTokens : null;
-  const titleParts = [
-    `agent ${run.ordinal + 1}`,
-    `status: ${run.status}`,
-    isSelected ? 'selected: chat shows this agent' : 'click to switch chat to this agent',
-    telemetry ? `provider: ${telemetry.provider}` : null,
-    telemetry ? `model: ${telemetry.model}` : null,
-    total !== null
-      ? `last turn: ${total} tokens · ${formatCost(telemetry!.estimatedCostUsd)}`
-      : null,
-  ].filter((p): p is string => p !== null);
-
-  const [draft, setDraft] = useState(run.name);
-  const [confirmingDelete, setConfirmingDelete] = useState(false);
-  useEffect(() => {
-    if (isEditing) {
-      setDraft(run.name);
-    }
-  }, [isEditing, run.name]);
-  useEffect(() => {
-    if (isEditing) {
-      setConfirmingDelete(false);
-    }
-  }, [isEditing]);
-
-  return (
-    <li
-      role={isEditing ? undefined : 'button'}
-      tabIndex={isEditing ? -1 : 0}
-      aria-pressed={isEditing ? undefined : isSelected}
-      onClick={isEditing ? undefined : onClick}
-      onDoubleClick={isEditing ? undefined : onRenameStart}
-      onKeyDown={(e) => {
-        if (isEditing) {
-          return;
-        }
-        if (e.key === 'Enter' || e.key === ' ') {
-          e.preventDefault();
-          onClick();
-        }
-      }}
-      className={cn(
-        'group rounded border transition-colors',
-        isEditing ? '' : 'cursor-pointer',
-        isSelected ? 'bg-elevated' : 'bg-muted/40 hover:bg-muted/60',
-        run.status === 'running'
-          ? 'border-info/60'
-          : agentHasUnread(run, isSelected && isTaskActive)
-            ? 'border-warning/70'
-            : isSelected
-              ? 'border-border'
-              : 'border-transparent',
-      )}
-    >
-      <div className="flex items-center gap-2 px-2 py-1.5" title={titleParts.join('\n')}>
-        <span
-          aria-hidden
-          className="w-4 shrink-0 text-right text-2xs tabular-nums text-muted-foreground/60"
-        >
-          {index + 1}.
-        </span>
-        <AgentKindChip
-          kind={kind}
-          title={`agent ${run.ordinal + 1}: ${AGENT_KIND_PALETTE[kind].label}`}
-        />
-        {isEditing ? (
-          <input
-            autoFocus
-            value={draft}
-            onChange={(e) => setDraft(e.target.value)}
-            onClick={(e) => e.stopPropagation()}
-            onDoubleClick={(e) => e.stopPropagation()}
-            onKeyDown={(e) => {
-              e.stopPropagation();
-              if (e.key === 'Enter') {
-                e.preventDefault();
-                onRenameCommit(draft);
-              } else if (e.key === 'Escape') {
-                e.preventDefault();
-                onRenameCancel();
-              }
-            }}
-            onBlur={() => onRenameCommit(draft)}
-            className="line-clamp-1 flex-1 rounded-full bg-background px-1.5 py-0.5 text-2xs font-medium text-foreground outline-none ring-1 ring-primary"
-            aria-label="rename agent"
-          />
-        ) : (
-          <span
-            className={cn(
-              'line-clamp-1 flex-1 text-left text-2xs font-medium',
-              isSelected ? 'text-foreground' : 'text-muted-foreground',
-            )}
-          >
-            {run.name}
-          </span>
-        )}
-        {!isEditing &&
-          (confirmingDelete ? (
-            <div className="flex shrink-0 items-center gap-1" onClick={(e) => e.stopPropagation()}>
-              <button
-                type="button"
-                onClick={() => setConfirmingDelete(false)}
-                className="rounded px-1 py-0.5 text-2xs text-muted-foreground transition-colors hover:bg-muted hover:text-foreground"
-                title="cancel"
-              >
-                Cancel
-              </button>
-              <button
-                type="button"
-                onClick={() => {
-                  setConfirmingDelete(false);
-                  onDelete();
-                }}
-                className="rounded px-1 py-0.5 text-2xs font-medium text-danger transition-colors hover:bg-danger/10"
-                title="confirm delete"
-              >
-                Delete
-              </button>
-            </div>
-          ) : (
-            <div className="flex shrink-0 items-center gap-1.5">
-              <ProviderGlyph provider={telemetry?.provider} />
-              <span className="text-2xs text-muted-foreground/70 group-hover:hidden">
-                <AgentLifetime run={run} />
-              </span>
-              <button
-                type="button"
-                onClick={(e) => {
-                  e.stopPropagation();
-                  setConfirmingDelete(true);
-                }}
-                className="hidden rounded p-0.5 text-muted-foreground/60 transition-colors group-hover:inline-flex hover:text-danger"
-                title="delete agent (double-click row to rename)"
-                aria-label="delete agent"
-              >
-                <Trash2 size={11} aria-hidden />
-              </button>
-            </div>
-          ))}
-      </div>
-      <div
-        className={cn(
-          'grid transition-[grid-template-rows] duration-200 ease-out',
-          isSelected ? 'grid-rows-[1fr]' : 'grid-rows-[0fr]',
-        )}
-      >
-        <div className="overflow-hidden">
-          <div className="flex flex-col gap-1 px-2 pb-1.5">
-            <AgentMetricsBlock
-              run={run}
-              telemetry={telemetry}
-              aggregate={aggregate}
-              turns={turns}
-              turnsLoading={turnsLoading}
-              variant="adhoc"
-            />
-            <ContextWindowBar telemetry={telemetry} aggregate={aggregate} />
-          </div>
-        </div>
-      </div>
-    </li>
-  );
-}
-
-function findContextWindow(model: string): number | null {
-  for (const cap of Object.values(PROVIDER_CAPABILITIES)) {
-    const m = cap.models.find((mm) => mm.id === model);
-    if (m) {
-      return m.contextWindow;
-    }
-  }
-  return null;
-}
-
-function AgentLifetime({ run }: { run: Agent }) {
-  const isLive = !!run.startedAt && !run.completedAt;
-  const now = useNow(5_000, isLive);
-  void now;
-
-  if (!run.startedAt) {
-    return (
-      <span
-        className="font-mono text-muted-foreground/60"
-        title="agent spawned but has not run yet"
-      >
-        0
-      </span>
-    );
-  }
-
-  const ageStr = formatRelativeDuration(run.startedAt, run.completedAt);
-  const tooltip = run.completedAt
-    ? `started ${run.startedAt}\ncompleted ${run.completedAt}\nworked ${ageStr}`
-    : `started ${run.startedAt}\nworking for ${ageStr}`;
-
-  return (
-    <span className="font-mono text-muted-foreground/80" title={tooltip}>
-      {ageStr}
-    </span>
-  );
-}
-
-function ContextWindowBar({
-  telemetry,
-  aggregate,
-}: {
-  telemetry: TelemetryRecord | null;
-  aggregate: AgentAggregate | null;
-}) {
-  if (!telemetry) {
-    return null;
-  }
-  const window = findContextWindow(telemetry.model);
-  if (!window) {
-    return null;
-  }
-  const cumulativeInput = aggregate?.inputTokens ?? telemetry.inputTokens;
-  const cumulativeOutput = aggregate?.outputTokens ?? telemetry.outputTokens;
-  const used = cumulativeInput + cumulativeOutput;
-  const pct = Math.min(1, used / window);
-  const barTone = (() => {
-    if (pct >= 0.9) {
-      return 'bg-danger';
-    }
-    if (pct >= 0.75) {
-      return 'bg-warning';
-    }
-    if (pct >= 0.5) {
-      return 'bg-info';
-    }
-    return 'bg-success';
-  })();
-  const iconTone = (() => {
-    if (pct >= 0.9) {
-      return 'text-danger';
-    }
-    if (pct >= 0.75) {
-      return 'text-warning';
-    }
-    if (pct >= 0.5) {
-      return 'text-info';
-    }
-    return 'text-success';
-  })();
-  const windowLabel = window >= 1_000_000 ? `${window / 1_000_000}M` : `${window / 1_000}k`;
-  const tooltip =
-    `context: ${used.toLocaleString()} / ${window.toLocaleString()} tokens (${Math.round(pct * 100)}%)\n` +
-    `cumulative input: ${cumulativeInput.toLocaleString()} · output: ${cumulativeOutput.toLocaleString()}`;
-  return (
-    <div className="flex flex-col gap-0.5" title={tooltip}>
-      <div className="flex items-center justify-between text-[9px] uppercase tracking-wide text-muted-foreground/60">
-        <span className={cn('flex items-center gap-0.5', iconTone)}>
-          <Gauge size={9} aria-hidden />
-          ctx
-        </span>
-        <span className="font-mono">
-          {formatTokens(used)} / {windowLabel} · {Math.round(pct * 100)}%
-        </span>
-      </div>
-      <div className="h-0.5 w-full overflow-hidden rounded-full bg-muted/60">
-        <div
-          className={cn('h-full rounded-full transition-all', barTone)}
-          style={{ width: `${pct * 100}%` }}
-        />
-      </div>
-    </div>
-  );
-}
-
-function WorkflowStartButton({
-  sessionId,
-  variant,
-}: {
-  sessionId: SessionId;
-  variant: 'empty' | 'attach';
-}) {
-  const onClick = () => {
-    window.dispatchEvent(
-      new CustomEvent('goodboy:open-workflow-builder', { detail: { sessionId } }),
-    );
-  };
-  return (
-    <div className={variant === 'empty' ? '' : 'mt-1.5'}>
-      <button
-        type="button"
-        onClick={onClick}
-        className="flex w-full items-center gap-2 rounded border border-dashed border-border-soft px-2 py-1.5 text-left text-xs text-muted-foreground transition-colors hover:border-border hover:bg-muted/50 hover:text-foreground"
-      >
-        <Plus size={13} aria-hidden className="shrink-0" />
-        <span className="min-w-0 truncate">
-          {variant === 'empty' ? 'Start a workflow' : 'Attach another workflow'}
-        </span>
-      </button>
-    </div>
-  );
-}
